src/test/java/me/star/webservice/web/service/PostsServiceTest.java
package me.star.webservice.web.service;

import me.star.webservice.domain.posts.Posts;
import me.star.webservice.domain.posts.PostsRepository;
import me.star.webservice.web.dto.PostsSaveRequestDto;
import org.junit.After;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.assertj.core.api.Assertions.assertThat;


@SpringBootTest
@RunWith(SpringRunner.class)
public class PostsServiceTest {

    @Autowired
    PostsRepository postsRepository;

    @Autowired
    PostsService postsService;

    @After
    public void cleanup () {
        postsRepository.deleteAll();
    }

    @Test
    public void 데이터_저장_잘되는가(){

        //given
        PostsSaveRequestDto dto = PostsSaveRequestDto.builder().
                title("test")
                .content("test conts")
                .author("mj")
                .build();

        //when
        postsService.save(dto);

        //then
        Posts posts = postsRepository.findAll().get(0);

        assertThat(posts.getTitle()).isEqualTo(dto.getTitle());
        assertThat(posts.getContent()).isEqualTo(dto.getContent());
        assertThat(posts.getAuthor()).isEqualTo(dto.getAuthor());


    }



}